deploy/aws/us-east-2/requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
flickrapi==2.4.0
idna==3.3
cffi==1.15.0
cryptography==36.0.2
PyJWT==2.3.0
oauthlib==3.2.0
psycopg2-binary==2.9.3
requests==2.27.1
requests-oauthlib==1.3.1
requests-toolbelt==0.9.1
six==1.16.0
urllib3==1.26.9
pycparser==2.21
